Split rules out into separate module
