+++ owners/test/index.test.js
@@ -20,7 +20,8 @@ const owners = require('..');
 const {Probot} = require('probot');
 const sinon = require('sinon');
 const {LocalRepository} = require('../src/local_repo');
-const {OwnersRule, OwnersParser} = require('../src/owners');
+const {OwnersParser} = require('../src/owners');
+const {OwnersRule} = require('../src/rules');
 
 const opened35 = require('./fixtures/actions/opened.35');
 const opened36 = require('./fixtures/actions/opened.36.author-is-owner');
