+++ owners/test/owners_check.test.js
@@ -22,7 +22,8 @@ const {
   CheckRunConclusion,
   OwnersCheck,
 } = require('../src/owners_check');
-const {OwnersRule, OwnersTree} = require('../src/owners');
+const {OwnersTree} = require('../src/owners');
+const {OwnersRule} = require('../src/rules');
 const {ReviewerSelection} = require('../src/reviewer_selection');
 
 describe('check run', () => {
